compiles... --kazu

+++ UserDev/SelectionTool/ERTool/Base/Env.h
@@ -23,10 +23,10 @@ namespace ertool {
   const double kINVALID_DOUBLE = kDOUBLE_MAX;
 
   /// Utility: invalid int value
-  const double kINVALID_INT = std::numeric_limits<int>::max();
+  const int kINVALID_INT = std::numeric_limits<int>::max();
 
   /// Utility: invalid size_t value
-  const double kINVALID_SIZE = std::numeric_limits<size_t>::max();
+  const size_t kINVALID_SIZE = std::numeric_limits<size_t>::max();
 
 }
